web: Refactor code structure
