+++ scripts/web/_env.js
@@ -1,0 +1,6 @@
+exports.IF = {
+  unit: 'rem',
+  planform: 'phone',
+  useRemote: false,
+  ctx: null
+}
